Updated
- Changed how resource groups are displayed, now they're "Name
(num/total)" format to quickly indicate the priority.
- Added TextAPI support along with a custom background for the text to
make it blend nicer with the vanilla UI.
- Added a menu to toggle the features in the mod on the fly.
- Added toggling of placement model transparency using
Shift+VoxelSettingsControl (or in the menu).
- Changed combination key for mount points to Ctrl+VoxelSettingsControl,
it's also in the menu.
- Fixed some instances where multiplied values ("Value (xMul)") weren't
actually multiplied by that value, now they should all be accurate to
what you experience in the world along with the paranthesis to indicate
that it's multiplied.
- Improved cargo mass estimate to only include components if the
inventory type is not defined, otherwise only include the type of cargo
the block accepts (like only ammo for turrets, only ore and ingots for
refinery, etc).
- Added some hardcoded values for vanilla upgrade modules (hardcoded
because of modAPI limitations).
- Reworded power requirement lines to be shorter.
- Reworded thrusters atmosphere influence lines.
- Improved thrusters flame distance and damage data (with some silly
complicated/hacky method), now it's in meters and in actual damage.
- Non-electricity thrustsers now accurately depict their power
requirements (which are none, they only need power to operate but do not
actually use power).
- Improved inventory information on blocks that have multiple
inventories (like refinery).
- Added some ways to get inventory information from blocks without
hardcoding it, the hardcoded values are still used as a fall back in
case the inventory data is not defined (Reminder: you can tell if it's
hardcoded by the small asterisk after the label).
- Reworded O2/s into oxygen/s as it's easier to read when next to zeres.
- Fixed turret, laser antenna and suspension wheel steer rotation speeds
being in the wrong notation, now they should be accurate as degrees per
second.