Updated
- Added air leak pathfinding feature for air vents:
--- Controlled from air vent's control panel via button.
--- Uses A* pathfinding algorithm to find the path towards the outside
of the ship starting from the air vent's position.
--- Gets processed in a background task so huge grids do not freeze your
game when finding air leaks.
--- Client side only; it's also a global thing, all air vents control
the same system, so you can start it from one air vent and stop it from
any other air vent; of course starting it uses the air vent's position.
--- Due to the pressurization code not being directly accessible it
needs to be copied into the mod, which means if they ever change it in
the game I need to copy it again, that is if they note the changes...

- Improved performance of all the generated text and using timed caches
to prevent frequent reprocessing of the same thing.
- Added "freeze in place" option (hotkey: Alt+VoxelSettings) which
freezes the placed block on the grid so you can inspect it from all
sides, really helpful with large blocks such as the refinery.
- Removed the Ship Drill radius stats because they were not actually
being updated from the definition; I've instead added hardcoded values
for the vanilla ones (marked with asterisk as usual).
- Added Remote Control power required and resource group info.
- Fixed doors showing as full airtight; also added a note there to
indicate that front&back faces get toggled airtightness.
- Added "first person view" and "can build" info for
cockpits/cryos/seats/RCs.
- Added "custom HUD name" to cockpits/seats/cryos.
- Added hardcoded vanilla Cryo Chamber power usage and sink group.
- Changed mod display to include the workshop ID.
- Changed orange to yellow for the textAPI entries which show up red in
the notification view mode.
- Lots of code rewrite and code organizing.