Bulk update
- Added aimed block text info for welders and grinders (both hand-held
and ship) - textAPI based for blending with vanilla UI and notification
based if textAPI is not installed.
- Added critical, functional and 'gets replaced on grind' colored lines
in the vanilla block info's components list - works regardless of
textAPI.
- Added ship tool aiming line to overlays mode (any except off) to
visualize how you need to aim to select a block.
- Added "/bi getblock [1~9]" command to equip a block from the world,
also added it in the build info menu.
- Updated TextAPI text and backgrounds to ignore postproecssing and
block flares - like vanilla HUD.
- Overlays also work for aimed block, some with small differences like
airtightness showing only if it's airtight in that build stage and doors
only if closed.
- Changed overlay for doors where it shows the blinking airtight face to
only show in airtightness overlay.
- Changed menu navigation to block rotation inputs.
- Changed "/bi" as the main command, with "/buildinfo" being the
obsolete one that only works for showing help screen.
- Fixed overlays not drawing at the correct position if in dynamic mode
and aiming at a voxel.
- Menu redesigned a bit in the textAPI mode.
- Many other code tweaks that aren't that relevant.
- Changed "/laserpower" command to "/bi laserpower"