New: Check game restart required
Game caches some things between world sessions which sneakily break things if those things change.

This feature also caches them into a file to persist between world loads + a file to identify when the game gets fully restarted to delete that data. Then when it detects a change that would probably break things, it alerts player to restart. Only if the offending mod is local it then provides more details.